refactor(jobs): local jobs listing moves into application/local_jobs
Suite RFC phase B. The on-disk jobs table (suite rows + unclaimed single
Attempts) is built in application/local_jobs/listing.py, which also owns
commands_for. viewer/jobs.py becomes the HTTP-facing forward layer and
keeps the projection helpers trials imports; LocalJobsCommands gains
list/get/overlay so 'ageval jobs list' goes through composition, and the
CLI no longer imports viewer.jobs (pinned by the composition-root test).
trials takes safe_id_segment from evidence.locators.
