+++ ARCHITECTURE.md
@@ -118,11 +118,11 @@ ageval/                              # GitHub: ZJU-REAL/ageval
 │   │   ├── lock.py                  # load_and_lock + capability/inject graph
 │   │   ├── run.py                   # mint identity → ctx → run_attempt
 │   │   ├── campaign.py
-│   │   ├── suite/                   # suite_run, fingerprint, suite_metrics
+│   │   ├── suite/                   # suite_run, fingerprint, suite_metrics, summary document
 │   │   ├── registry_ops/            # results / publish / login / org / list
 │   │   ├── plugin_ops/
 │   │   ├── agent_ops/               # --agent (+ optional --model) projects into profiles
-│   │   └── local_jobs/              # local Job delete
+│   │   └── local_jobs/              # local Job list / get / delete
 │   ├── attempt/                     # deep module: one Attempt’s visible pipeline
 │   │   ├── __init__.py              # run_attempt five-phase lines
 │   │   ├── ctx.py
