+++ src/ageval/viewer/trials/fs.py
@@ -8,7 +8,8 @@ from pathlib import Path
 from typing import Any
 
 from ageval.config.errors import ConfigError
-from ageval.viewer.jobs import get_job, safe_id_segment
+from ageval.evidence.locators import safe_id_segment
+from ageval.viewer.jobs import get_job
 from ageval.viewer.trials.constants import MAX_FILE_BYTES, MAX_TREE_ENTRIES, TEXT_SUFFIXES
 from ageval.viewer.trials.paths import (
     _read_json_object,
